requirements.txt
# Email Outreach Automation System Requirements
# Python 3.9+ required

# Core Google API libraries
google-api-python-client>=2.170.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# Web scraping and HTTP requests
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

# Optional: Better HTTP client
httpx>=0.24.0

# Development and type checking (optional)
mypy>=1.5.0
black>=23.0.0
flake8>=6.0.0

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0